requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
Django==1.11.11
django-haystack==2.8.1
django-redis==4.10.0
django-tinymce==2.7.0
dnspython==1.15.0
eventlet==0.24.1
greenlet==0.4.15
idna==2.7
itsdangerous==1.1.0
jieba==0.39
kombu==4.2.1
monotonic==1.5
mutagen==1.41.1
Pillow==5.3.0
pycparser==2.19
pycryptodomex==3.7.2
PyMySQL==0.9.2
python-alipay-sdk==1.8.0
pytz==2018.7
redis==2.10.6
requests==2.20.1
six==1.11.0
urllib3==1.24.1
vine==1.1.4
Whoosh==2.7.4
